+++ interfaces/src/main/java/ar/edu/itba/paw/services/UserBookingService.java
@@ -7,12 +7,12 @@ import java.util.List;
 public interface UserBookingService<T extends User> {
     List<Booking> getBookings(T user, BookingState state, int page);
 
-    long getBookingCount(T user, BookingState state);
+    int getBookingCount(T user, BookingState state);
 
     default int getBookingPages(T user, BookingState state) {
-        long totalBookings = this.getBookingCount(user, state);
+        int totalBookings = this.getBookingCount(user, state);
         return (int) Math.ceil((double) totalBookings / Pagination.BOOKINGS_PAGE_SIZE);
     }
 
-    Booking cancelBooking(long bookingId, T user);
+    Booking cancelBooking(int bookingId, T user);
 }
