+++ persistence/src/main/java/ar/edu/itba/paw/persistence/ImageJpaDao.java
@@ -15,11 +15,11 @@ public class ImageJpaDao implements ImageDao {
     private EntityManager em;
 
     @Override
-    public Image getImage(long id) {
+    public Image getImage(int id) {
         return em.find(Image.class, id);
     }
 
-    private long uploadImage(String fileName, byte[] imgData) {
+    private int uploadImage(String fileName, byte[] imgData) {
         Image image = new Image();
         image.setFileName(fileName);
         image.setData(imgData);
@@ -29,24 +29,24 @@ public class ImageJpaDao implements ImageDao {
     }
 
     @Override
-    public long uploadPop(byte[] bin, String fileName, Booking booking) {
-        long imageId = uploadImage(fileName, bin);
+    public int uploadPop(byte[] bin, String fileName, Booking booking) {
+        int imageId = uploadImage(fileName, bin);
         booking.setPop(imageId);
         em.merge(booking);
         return imageId;
     }
 
     @Override
-    public long uploadVehicleImage(byte[] bin, String fileName, Vehicle vehicle) {
-        long imageId = uploadImage(fileName, bin);
+    public int uploadVehicleImage(byte[] bin, String fileName, Vehicle vehicle) {
+        int imageId = uploadImage(fileName, bin);
         vehicle.setImgId(imageId);
         em.merge(vehicle);
         return imageId;
     }
 
     @Override
-    public long uploadPfp(byte[] bin, String fileName, User user) {
-        long imageId = uploadImage(fileName, bin);
+    public int uploadPfp(byte[] bin, String fileName, User user) {
+        int imageId = uploadImage(fileName, bin);
         user.setPfp(imageId);
         em.merge(user);
         return imageId;
